refactor(default): remove the "newline-per-chained-call" rule
Sometimes you want to chain lot of calls, e.g. in your tests:
expect(myObj).to.be.an("object").that.is.not.null;

whereas other times you clearly want to add "space":
myPromise
.then(result => console.log(result))
.catch(err => console.error(err));

Because the only option was to set the "minimum" number of calls
before adding a new line this rule doesn't offer much flexibility
so i've removed it!